chunk out js
